reshenie kubov

+++ seminar3/03/Program.cs
@@ -1,6 +1,6 @@
 ﻿/*Напишите программу, которая принимает на вход число (N) и выдаёт таблицу кубов чисел от 1 до N.*/
 Console.Write("Введите число N: ");
-double N = Convert.ToDouble(Console.ReadLine());
+int N = Convert.ToInt32(Console.ReadLine());
 if (N<0) N=-N;
 int index = 1;
 while (index<=N)
